refactor(pdf): isolate equation source crop planning
